+++ Documentation/ReleaseNotes.md
@@ -1,4 +1,15 @@
 # Release notes
+## 2.0.0
+*Breaking changes only for Index page functionality.*
+
+**New**
+* Added functionality to show actual configuration values (filtered) on index page and their source.
+  * \+ Configuration value obfuscation helpers.
+* Added functionality to include custom content (file) into index page.
+
+**Improved**
+* Setting up Index page in fluent manner. Now only dealing with `IndexPage` object itself.
+
 ## 1.1.0
 **New**
 * Added options setting for ErrorHandler. This allows additionally to filter out elements in stacktrace frames.